ATS/requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
mlflow
spacy
sentence_transformers
pdfplumber